sending hints to lobby closes #73

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/service/GameService.java
@@ -53,7 +53,7 @@ public class GameService {
 
         GameRepository.addGame(lobby.getLobbyId(), game);
 
-        lobby.setCurrentGameId(game.getGameId().longValue());
+        lobby.setCurrentGameId(game.getGameId());
         lobby = this.lobbyRepository.save(lobby);
         this.lobbyRepository.flush();
